pyproject.toml
[project]
name = "airflow-aws-kickstart"
version = "0.1.0"
description = "Sample Apache Airflow 3 Project and DAG to kickstart your Airflow + AWS data engineering journey"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "apache-airflow>=3.0.6",
    "boto3>=1.40.37",
]
